Add Start component, update Home
